+++ javascripts/init/init.js
@@ -1,6 +1,29 @@
 // Variables
-var clickConfigStart = 0;
-var EtapeConfig = 0;
+var clickConfigStart = 0,
+    EtapeConfig = 0,
+    progressBar = 252,
+    color = '',
+    DB_HOST = '',
+    DB_PORT = '',
+    DB_NAME = '',
+    DB_USER = '',
+    DB_PASSWORD = '',
+    USER_ADMIN = '',
+    USER_ADMIN_MDP = '';
+
+function ResetVar() {
+    clickConfigStart = 0,
+        EtapeConfig = 0,
+        progressBar = 252,
+        color = '',
+        DB_HOST = '',
+        DB_PORT = '',
+        DB_NAME = '',
+        DB_USER = '',
+        DB_PASSWORD = '',
+        USER_ADMIN = '',
+        USER_ADMIN_MDP = '';
+}
 
 // CHANGEMENT DE PAGE
 $('.custom-button').click(function (e) {
@@ -10,9 +33,10 @@ $('.custom-button').click(function (e) {
         stepBlock = $('.step-content'),
         stepMenu = $('.step');
 
-    if (step == "" || step == undefined || step == null) {
+    if (step == "" || step == undefined || step == null || $(this).hasClass('disable')) {
         return;
     }
+
     // Retrait des class "active"
     stepBlock.removeClass('active');
     stepMenu.removeClass('active');
@@ -22,20 +46,28 @@ $('.custom-button').click(function (e) {
     $('#step' + step).addClass('active');
 })
 
-function LoadBar(etape, pixel) {
-    console.log('click');
-    var nouvelleOmbre = 'inset ' + pixel + 'px 0px 0px 0px rgba(202, 131, 0, 0.6)';
+// Bar de chargement
+function LoadBar(etape, pixel, resultat) {
+    if (resultat == "succes") {
+        color = "rgb(32, 160, 0)";
+    } else if (resultat == "error") {
+        color = "rgb(255, 0, 0)";
+    } else {
+        color = "rgba(202, 131, 0, 0.6)";
+    }
+
+    var nouvelleOmbre = 'inset ' + pixel + 'px 0px 0px 0px ' + color;
     // Appliquez la nouvelle valeur de la propriété box-shadow à l'élément
     $('#step3-etp' + etape).css('box-shadow', nouvelleOmbre);
 }
 
 function ChangeIcoText(etape, ico, text) {
     switch (ico) {
-        case "spinner":
+        case "load":
             ico = "fa-spinner"
             break;
 
-        case "good":
+        case "succes":
             ico = "fa-circle-check"
             break;
 
@@ -57,58 +89,139 @@ function ChangeIcoText(etape, ico, text) {
     $('#step3-etp' + etape).children('span').text(text);
 }
 
-function StartConfig() {
-    clickConfigStart = 1;
-    EtapeConfig = EtapeConfig++;
+function StartConfig(step, textStats) {
     if (clickConfigStart == 0) {
+        EtapeConfig++;
+        // Blocage du clique
+        clickConfigStart = 1;
+        // Définission des variables définitives
+        DB_HOST = $('#DB_HOST').val(),
+            DB_PORT = $('#DB_PORT').val(),
+            DB_NAME = $('#DB_NAME').val(),
+            DB_USER = $('#DB_USER').val(),
+            DB_PASSWORD = $('#DB_PASSWORD').val(),
+            USER_ADMIN = $('#USER_ADMIN').val(),
+            USER_ADMIN_MDP = $('#USER_ADMIN_MDP').val();
+
+        // Modification du bouton
         $('#start-check').text('Patientez');
         $('#start-check').addClass('disable');
+
+        // Définission de la nouvelle étape
+        if (!step) {
+            step = "step3-1.1";
+            textStats = "Tentative de connexion";
+        }
+
+        // Utilisation de la progressBar
+        LoadBar(EtapeConfig, progressBar / 2);
+        ChangeIcoText(EtapeConfig, 'load', textStats);
+
+        // Lancement de la fonction
+        setTimeout(() => {
+            CallConfig(EtapeConfig, step);
+        }, 2000);
+
     }
-    CallConfig();
 }
 
-function ChangeConfig() {
+function CallConfig(EtapeConfig, step) {
+    // Préparation de la requête Ajax
+    $.ajax({
+        url: './init/' + step, // Remplacer "step3..." par la variable "step"
+        type: 'POST',
+        data: {
+            DB_HOST: DB_HOST,
+            DB_PORT: DB_PORT,
+            DB_NAME: DB_NAME,
+            DB_USER: DB_USER,
+            DB_PASSWORD: encodeURIComponent(DB_PASSWORD),
+            USER_ADMIN: USER_ADMIN,
+            USER_ADMIN_MDP: encodeURIComponent(USER_ADMIN_MDP)
+        },
+        success: function (response) {
+            // PARSE De la réponse
+            response = JSON.parse(response);
+            console.log(response);
 
-    switch (key) {
-        case 'check-bdd-connect':
-            // Page 3, Etape 1, Première relance
-            FunctResponse = CallConfig('3-1.1');
+            if (response.nextStep === "STOP") {
+                // Progression FULL
+                LoadBar(EtapeConfig, '252', response.resultat);
+                // Changement du logo et texte
+                ChangeIcoText(EtapeConfig, response.resultat, response.message);
 
-            // Fonction d'action 
-            function name(params) {
-                
+                $('#start-check').text('Patientez ....');
+
+                // Changement de la page ....
+                setTimeout(() => {
+                    let stepBlock = $('.step-content'),
+                        stepMenu = $('.step');
+
+                    // Retrait des class "active"
+                    stepBlock.removeClass('active');
+                    stepMenu.removeClass('active');
+
+                    // Ajout de la class "active" au nouvelle div
+                    $('#menu-step4').addClass('active');
+                    $('#step4').addClass('active');
+                }, 1000);
+
+                // Changement de la page ....
+                setTimeout(() => {
+                    window.location.href = window.location.href;
+                }, 3000);
+                // Déblocage du bouton
+                clickConfigStart = 0;
+                return
             }
 
-            // Réponse en json avec ok, warning, ou error.
-            // ok = Passage à l'étape suivante
-            // warning = Erreur non bloquante. Potentiellement un skip (Ex: BDD déjà créé)
-            // error = Erreur bloquante, arrêt du script et des étape.
-            break;
-    
-        default:
-            break;
-    }
+            if (response.resultat == "succes" || response.resultat == "warning") {
+                // Progression FULL
+                LoadBar(EtapeConfig, '252', response.resultat);
+                // Changement du logo et texte
+                ChangeIcoText(EtapeConfig, response.resultat, response.message);
+                // Changement d'étape
+                EtapeConfig++;
+                clickConfigStart = 0;
+                setTimeout(() => {
+                    StartConfig(response.nextStep, response.textStats);
+                }, 2000);
 
+            } else if (response.resultat == "error") {
+                // Progression FULL
+                LoadBar(EtapeConfig, '252', response.resultat);
+                // Changement du logo et texte
+                ChangeIcoText(EtapeConfig, response.resultat, response.message);
+                // Rétablissement du bouton
+                $('#start-check').text('Lancement');
+                $('#start-check').removeClass('disable');
+                $('.cancel').removeClass('disable');
 
-    LoadBar('125', '')
-    ChangeIcoText(etape, 'spinner', 'TEST');
-}
-
-function CallConfig() {
-    $.ajax({
-        url: './init/step3-1.1', // Remplacez par votre script côté serveur qui gère l'ajout
-        type: 'POST',
-    //     data: {
-    //         nomComplet: nomComplet.val(),
-    //         identifiant: identifiant.val(),
-    //         email: email.val(),
-    //         account_administrator: account_administrator.val()
-    // },
-        success: function (response) {
-            console.log(response);
+                // Reset STEP
+                ResetVar();
+                // Déblocage du bouton
+                clickConfigStart = 0;
+                return;
+            }
         },
         error: function (xhr, status, error) {
-            console.error(xhr.responseText);
+            // console.error(xhr.responseText);
+
+            // Rétablissement du bouton
+            $('#start-check').text('Lancement');
+            $('#start-check').removeClass('disable');
+            $('.cancel').removeClass('disable');
+
+            // Déblocage du bouton
+            clickConfigStart = 0;
+
+            // Blocage de la barre
+            LoadBar(EtapeConfig, '252', 'error');
+            // Changement du logo et texte
+            ChangeIcoText(EtapeConfig, 'error', 'Erreur interne #1000');
+
+            // Reset STEP
+            ResetVar();
         }
     });
 }