+++ CHANGELOG.md
@@ -5,6 +5,27 @@
 - Ajout d'un btn menu dans le header
 - Ajout d'un nouveau menu à gauche
 - Déportation des options de l'ancien menu vers le nouveau
+- Création de la nouvelle page de configuartion
+- Ajout d'un système de page par étape
+- Création du nouveau formulaire
+- Création d'un tableau avec les étapes
+- Modification du routeur pour la page d'initialisation (Pas de changement de fonctionnement du routeur)
+- Page d'initialisation fonctionnel mais BETA
+- Modification du formulaire de connexion (Changement de design)
+- UPDATE de la table Nino (DB06)
+- Correction de bogues mineurs dans l'édition de Nino
+- Mise en place du nouveau lecteur
+- ProgressBar de chargement de vidéo
+- ProgressBar de timeline
+- Bouton Pause/Play
+- Bouton +10 secondes et -10 secondes
+- Bouton du mode cinéma
+- Activation du FullScreen (Commande de base du navigateur)
+- Bouton de supression d'une route
+
+
+
+========== ### V4  ==========
 
 ### 4.7.7
 - Ajout de la miniature sur la preview (BETA)